% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premixed.R
\name{coef.premixed}
\alias{coef.premixed}
\title{Return estimated fixed-effects coefficients from a mixed-effects prediction
rule ensemble}
\usage{
\method{coef}{premixed}(object, ...)
}
\arguments{
\item{object}{an object of class 'premixed'.}

\item{...}{further arguments to be passed to \code{\link[pre]{coef.pre}}.}
}
\description{
\code{coef.premixed} returns the estimated fixed-effects coefficients from 
a mixed-effects prediction rule ensemble
}
