% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premixed.R
\name{complexity}
\alias{complexity}
\title{Calculates complexity of a prediction rule ensemble.}
\usage{
complexity(object, penalty.par.val = "lambda.1se", ...)
}
\arguments{
\item{object}{an object of class 'pre' or premixed'.}

\item{penalty.par.val}{As usual.}

\item{...}{not currently used.}
}
\value{
Returns a vector with the total number of variables in the ensemble,
the total number of terms (i.e., baselearners with a non-zero coefficient),
and the mean number of variables per term) in the ensemble.
}
\description{
\code{complexity} returns the complexity (total number of variables in the 
ensemble, number of terms and mean number of variables per term) of a
prediction rule ensembles (i.e., an object of class 'pre' or 'premixed').
}
\examples{
\donttest{
set.seed(42)
airq <- airquality[complete.cases(airquality),]
airq.ens1 <- premixed(Ozone ~ 1 | Month | Solar.R + Wind + Temp + Day, data = airq, ntrees = 10)
complexity(airq.ens1)
}

}
