% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premixed.R
\name{predict.premixed}
\alias{predict.premixed}
\title{Return predicted values based on a mixed-effects prediction rule ensemble.}
\usage{
\method{predict}{premixed}(object, offset = 0, ...)
}
\arguments{
\item{object}{an object of class 'premixed'.}

\item{offset}{Offset value to be used in generating predictions. Note that
the predictions are returned for the fixed effects only.}

\item{...}{further arguments to be passed to \code{\link[pre]{predict.pre}}}
}
\description{
\code{predict.premixed} returns predicted values based on the estimated
fixed effects of a a mixed-effects prediction rule ensemble (random 
effects are NOT included in the predictions (yet)).
}
\examples{
\donttest{
## Employ glmertree for rule induction and iterate between lasso for
## estimating fixed effects and ridge for estimating random effects:
set.seed(42)
airq <- airquality[complete.cases(airquality),]
airq.ens1 <- premixed(Ozone ~ 1 | Month | Solar.R + Wind + Temp + Day, data = airq, ntrees = 10)
predict(airq.ens1, newdata = airq)
}
}
