% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premixed.R
\name{ranef.premixed}
\alias{ranef.premixed}
\title{Return predicted random-effects coefficients from a mixed-effects prediction 
rule ensemble}
\usage{
\method{ranef}{premixed}(object, ...)
}
\arguments{
\item{object}{an object of class'premixed'.}

\item{...}{currently not used.}
}
\description{
\code{ranef.premixed} returns the predicted random-effects coefficients from a 
mixed-effects prediction rule ensemble.
}
