% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/premixed.R
\name{premixed}
\alias{premixed}
\title{Fitting a prediction rule ensemble with clustered data}
\usage{
premixed(formula, cluster = NULL, data, penalty.par.val = "lambda.min",
  learnrate = 0, use.grad = FALSE, conv.thresh = 0.5,
  family = "gaussian", max.iter = 1000, ...)
}
\arguments{
\item{formula}{a Formula object with three-part right-hand side like 
y ~ 1 | cluster | x1 + x2 + x3}

\item{cluster}{optional string supplying the name of the cluster indicator. If specified,
\code{formula} should not involve random effects (e.g., 
\code{y ~ x1+ x2 + x3}). If \code{cluster} is specified, random effects will
not be estimated during tree induction. This will substantially speed up 
computations, but may yield a less accurate model, depending on the magnitude
of the random effects.}

\item{data}{a dataframe containing the variables in the model}

\item{penalty.par.val}{as usual.}

\item{learnrate}{as usual.}

\item{use.grad}{as usual.}

\item{conv.thresh}{numeric vector of length 1, specifies the convergence
criterion. The algorithm converges if the maximum absolute difference in 
the random-effects coefficients from one iteration is smaller than 
\code{conv.thresh}.}

\item{family}{as usual.}

\item{max.iter}{numeric vector of length 1. Maximum number of iterations
performed to re-estimate fixed and random effects parameters.}

\item{...}{further arguments to be passed to \code{\link[pre]{pre}}.}
}
\value{
An object of class 'premixed'.
}
\description{
Experimental function, use at own risk. Estimates rules by 
using function glmertree(). Estimates random- and fixed-effects coefficients
by iterating between lasso estimation of the fixed-effects (i.e., rules 
and/or linear terms) and ridge estimation of the random-effects predictions.
The model involves a random intercept only.
}
\details{
\code{premixed} is a wrapper for fitting a PRe with clustered data
}
\examples{
\donttest{set.seed(42)
}
}
